Replace Alias Method Chain With Prepend for SchemaDumper
