+++ lib/activerecord_multi_tenant.rb
@@ -12,3 +12,4 @@ require_relative 'activerecord-multi-tenant/table_node'
 require_relative 'activerecord-multi-tenant/version'
 require_relative 'activerecord-multi-tenant/habtm'
 require_relative 'activerecord-multi-tenant/relation_extension'
+require_relative 'activerecord-multi-tenant/schema_dumper'
